# product/activation-server/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "U-Claw Activation API",
    "version": "1.0.0"
  },
  "servers": [{ "url": "https://api.u-claw.org" }],
  "paths": {
...
    "/internal/v1/inventory": {
      "post": {
        "operationId": "adminGenerateInventory", "security": [{ "AdminBearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminGenerateRequest" } } } },
        "responses": { "201": { "description": "Generated inventory credentials", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/AdminInventorySecret" } } } } }, "400": { "$ref": "#/components/responses/AdminError" }, "401": { "$ref": "#/components/responses/AdminError" }, "503": { "$ref": "#/components/responses/AdminError" } }
      }
    },
    "/internal/v1/inventory/import": {
      "post": {
        "operationId": "adminImportInventory", "security": [{ "AdminBearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminImportRequest" } } } },
        "responses": { "201": { "description": "Imported inventory", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/AdminInventorySummary" } } } } }, "400": { "$ref": "#/components/responses/AdminError" }, "401": { "$ref": "#/components/responses/AdminError" }, "503": { "$ref": "#/components/responses/AdminError" } }
      }
    },
    "/internal/v1/inventory/{id}": {
      "get": {
        "operationId": "adminShowInventory", "security": [{ "AdminBearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/AdminId" }],
        "responses": { "200": { "description": "Inventory summary", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminInventorySummary" } } } }, "401": { "$ref": "#/components/responses/AdminError" }, "503": { "$ref": "#/components/responses/AdminError" } }
      }
    },
...
    "/internal/v1/new-api-bindings/{deviceId}/balance-status": {
      "patch": {
        "operationId": "adminUpdateBalanceStatus", "security": [{ "AdminBearerAuth": [] }], "parameters": [{ "name": "deviceId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Identifier" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminBalanceStatusRequest" } } } },
        "responses": { "200": { "description": "Binding marked configured", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminInventorySummary" } } } }, "400": { "$ref": "#/components/responses/AdminError" }, "401": { "$ref": "#/components/responses/AdminError" }, "503": { "$ref": "#/components/responses/AdminError" } }
      }
    },
    "/internal/v1/new-api-bindings/{inventoryId}": {
      "get": {
        "operationId": "adminShowNewAPIMapping", "security": [{ "AdminBearerAuth": [] }],
        "parameters": [{ "name": "inventoryId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Identifier" } }],
        "responses": { "200": { "description": "Redacted New API mapping", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminMappingSummary" } } } }, "400": { "$ref": "#/components/responses/AdminError" }, "401": { "$ref": "#/components/responses/AdminError" }, "503": { "$ref": "#/components/responses/AdminError" } }
      }
    },
...
    "/internal/v1/audit": {
      "get": {
        "operationId": "adminQueryAudit", "security": [{ "AdminBearerAuth": [] }],
        "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 500, "default": 100 } }, { "name": "before", "in": "query", "description": "Opaque base64url cursor containing the preceding event timestamp and UUID", "schema": { "type": "string", "minLength": 1, "maxLength": 512 } }],
        "responses": { "200": { "description": "Audit events", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminAuditPage" } } } }, "400": { "$ref": "#/components/responses/AdminError" }, "401": { "$ref": "#/components/responses/AdminError" }, "503": { "$ref": "#/components/responses/AdminError" } }
      }
    },
    "/internal/v1/releases/publish": {
      "post": {
        "operationId": "adminPublishRelease", "security": [{ "AdminBearerAuth": [] }],
        "description": "Atomically move production.previousStable to current and production.current to a fully read-back verified higher releaseSequence.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminReleasePublishRequest" } } } },
        "responses": { "200": { "description": "Production release slots", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductionReleaseSlots" } } } }, "400": { "$ref": "#/components/responses/AdminError" }, "401": { "$ref": "#/components/responses/AdminError" }, "503": { "$ref": "#/components/responses/AdminError" } }
      }
    },
    "/internal/v1/releases/forward-rollback": {
      "post": {
        "operationId": "adminForwardRollbackRelease", "security": [{ "AdminBearerAuth": [] }],
        "description": "Publish previousStable content under a new higher releaseSequence and mark the failed current release withdrawn.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminReleaseRollbackRequest" } } } },
        "responses": { "200": { "description": "Production release slots", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductionReleaseSlots" } } } }, "400": { "$ref": "#/components/responses/AdminError" }, "401": { "$ref": "#/components/responses/AdminError" }, "503": { "$ref": "#/components/responses/AdminError" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer" },
      "AdminBearerAuth": { "type": "http", "scheme": "bearer", "description": "Independent admin secret" }
    },
...
    "parameters": {
...
      "AdminId": { "name": "id", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Identifier" } }
...
    },
    "responses": {
...
      "AdminError": { "description": "Redacted admin error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminError" } } } },
...
    },
    "schemas": {
      "Identifier": {
        "type": "string",
        "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{2,127}$"
      },
      "ActivationCode": { "type": "string", "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$" },
...
      "AdminGenerateRequest": { "type": "object", "additionalProperties": false, "required": ["operatorId", "requestId", "idempotencyKey", "reason", "count"], "properties": { "operatorId": { "$ref": "#/components/schemas/Identifier" }, "requestId": { "$ref": "#/components/schemas/Identifier" }, "idempotencyKey": { "type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$" }, "reason": { "type": "string", "minLength": 3, "maxLength": 512 }, "count": { "type": "integer", "minimum": 1, "maximum": 10000 } } },
      "AdminBalanceStatusRequest": { "type": "object", "additionalProperties": false, "required": ["operatorId", "requestId", "idempotencyKey", "reason", "balanceStatus"], "properties": { "operatorId": { "$ref": "#/components/schemas/Identifier" }, "requestId": { "$ref": "#/components/schemas/Identifier" }, "idempotencyKey": { "type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$" }, "reason": { "type": "string", "minLength": 3, "maxLength": 512 }, "balanceStatus": { "const": "configured" } } },
      "AdminImportRequest": { "type": "object", "additionalProperties": false, "required": ["operatorId", "requestId", "idempotencyKey", "reason", "records"], "properties": { "operatorId": { "$ref": "#/components/schemas/Identifier" }, "requestId": { "$ref": "#/components/schemas/Identifier" }, "idempotencyKey": { "type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$" }, "reason": { "type": "string", "minLength": 3, "maxLength": 512 }, "records": { "type": "array", "minItems": 1, "maxItems": 10000, "items": { "type": "object", "additionalProperties": false, "required": ["username", "activationCode", "newApiUserId", "newApiUsername", "policyDigest"], "properties": { "username": { "$ref": "#/components/schemas/Identifier" }, "activationCode": { "$ref": "#/components/schemas/ActivationCode" }, "newApiUserId": { "$ref": "#/components/schemas/Identifier" }, "newApiUsername": { "$ref": "#/components/schemas/Identifier" }, "policyDigest": { "type": "string", "pattern": "^[a-f0-9]{64}$" } } } } } },
      "AdminInventorySecret": { "type": "object", "additionalProperties": false, "required": ["inventoryId", "username", "activationCode"], "properties": { "inventoryId": { "$ref": "#/components/schemas/Identifier" }, "username": { "type": "string" }, "activationCode": { "$ref": "#/components/schemas/ActivationCode" } } },
      "AdminInventorySummary": { "type": "object", "additionalProperties": false, "required": ["inventoryId", "username", "status", "newApiSetupStatus", "deviceId", "licenseId"], "properties": { "inventoryId": { "$ref": "#/components/schemas/Identifier" }, "username": { "$ref": "#/components/schemas/Identifier" }, "status": { "enum": ["prepared", "binding", "active", "revoked"] }, "newApiSetupStatus": { "enum": ["pending", "configured", "suspended"] }, "deviceId": { "anyOf": [{ "$ref": "#/components/schemas/Identifier" }, { "type": "null" }] }, "licenseId": { "anyOf": [{ "$ref": "#/components/schemas/Identifier" }, { "type": "null" }] } } },
...
      "AdminMappingSummary": { "type": "object", "additionalProperties": false, "required": ["inventoryId", "newApiUserId", "newApiUsername", "baseUrlHost", "defaultModel", "allowedModels", "requestsPerMinute", "concurrentRequests", "keyVersion", "status"], "properties": { "inventoryId": { "$ref": "#/components/schemas/Identifier" }, "newApiUserId": { "$ref": "#/components/schemas/Identifier" }, "newApiUsername": { "$ref": "#/components/schemas/Identifier" }, "baseUrlHost": { "type": "string" }, "defaultModel": { "type": "string" }, "allowedModels": { "type": "array", "minItems": 1, "uniqueItems": true, "items": { "type": "string", "minLength": 1 } }, "requestsPerMinute": { "type": "integer", "minimum": 1, "maximum": 6000 }, "concurrentRequests": { "type": "integer", "minimum": 1, "maximum": 100 }, "keyVersion": { "$ref": "#/components/schemas/Identifier" }, "status": { "type": "string" }, "updatedAt": { "type": "string", "format": "date-time" } } },
...
      "AdminAuditEvent": { "type": "object", "additionalProperties": false, "required": ["eventId", "actorId", "action", "outcome", "inventoryId", "deviceId", "licenseId", "requestId", "reason", "idempotencyKey", "createdAt"], "properties": { "eventId": { "$ref": "#/components/schemas/Identifier" }, "actorId": { "type": "string" }, "action": { "type": "string" }, "outcome": { "enum": ["succeeded", "failed"] }, "inventoryId": { "anyOf": [{ "$ref": "#/components/schemas/Identifier" }, { "type": "null" }] }, "deviceId": { "anyOf": [{ "$ref": "#/components/schemas/Identifier" }, { "type": "null" }] }, "licenseId": { "anyOf": [{ "$ref": "#/components/schemas/Identifier" }, { "type": "null" }] }, "requestId": { "$ref": "#/components/schemas/Identifier" }, "reason": { "type": ["string", "null"] }, "idempotencyKey": { "anyOf": [{ "type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$" }, { "type": "null" }] }, "createdAt": { "type": "string", "format": "date-time" } } },
      "AdminAuditPage": { "type": "object", "additionalProperties": false, "required": ["items", "nextBefore"], "properties": { "items": { "type": "array", "maxItems": 500, "items": { "$ref": "#/components/schemas/AdminAuditEvent" } }, "nextBefore": { "type": ["string", "null"], "maxLength": 512 } } },
      "AdminError": { "type": "object", "additionalProperties": false, "required": ["code"], "properties": { "code": { "enum": ["ADMIN_AUTHENTICATION_FAILED", "ADMIN_INVALID", "ADMIN_SERVICE_UNAVAILABLE"] } } },
...
      "ReleaseId": { "type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$" },
      "SHA256": { "type": "string", "pattern": "^[a-f0-9]{64}$" },
...
      "AdminReleasePublishRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["operatorId", "requestId", "idempotencyKey", "reason", "releaseSequence", "releaseId", "contentVersion", "manifestUrl", "manifestSha256", "manifestReadbackVerified", "cdnAvailable"],
        "properties": {
          "operatorId": { "$ref": "#/components/schemas/Identifier" }, "requestId": { "$ref": "#/components/schemas/Identifier" }, "idempotencyKey": { "$ref": "#/components/schemas/Identifier" }, "reason": { "type": "string", "minLength": 1, "maxLength": 512 },
          "releaseSequence": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 }, "releaseId": { "$ref": "#/components/schemas/ReleaseId" }, "contentVersion": { "type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9.+_-]{0,63}$" }, "manifestUrl": { "type": "string", "format": "uri", "pattern": "^https://" }, "manifestSha256": { "$ref": "#/components/schemas/SHA256" }, "manifestReadbackVerified": { "const": true }, "cdnAvailable": { "const": true }
        }
      },
      "AdminReleaseRollbackRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["operatorId", "requestId", "idempotencyKey", "reason", "releaseSequence", "releaseId", "manifestUrl", "manifestSha256", "manifestReadbackVerified", "cdnAvailable"],
        "properties": {
          "operatorId": { "$ref": "#/components/schemas/Identifier" }, "requestId": { "$ref": "#/components/schemas/Identifier" }, "idempotencyKey": { "$ref": "#/components/schemas/Identifier" }, "reason": { "type": "string", "minLength": 1, "maxLength": 512 },
          "releaseSequence": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 }, "releaseId": { "$ref": "#/components/schemas/ReleaseId" }, "manifestUrl": { "type": "string", "format": "uri", "pattern": "^https://" }, "manifestSha256": { "$ref": "#/components/schemas/SHA256" }, "manifestReadbackVerified": { "const": true }, "cdnAvailable": { "const": true }
        }
      },
      "ProductionReleaseSlot": {
        "type": "object", "additionalProperties": false, "required": ["releaseSequence", "releaseId", "contentVersion", "reason", "status"],
        "properties": { "releaseSequence": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 }, "releaseId": { "$ref": "#/components/schemas/ReleaseId" }, "contentVersion": { "type": "string" }, "reason": { "enum": ["release", "rollback"] }, "status": { "enum": ["current", "stable"] } }
      },
      "ProductionReleaseSlots": {
        "type": "object", "additionalProperties": false, "required": ["policyEpoch", "current", "previousStable"],
        "properties": { "policyEpoch": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 }, "current": { "$ref": "#/components/schemas/ProductionReleaseSlot" }, "previousStable": { "oneOf": [{ "$ref": "#/components/schemas/ProductionReleaseSlot" }, { "type": "null" }] } }
      },
...
    }
  }
}